requirements.txt
pydantic>=2.7
python-dotenv>=1.0.1
requests>=2.32
pandas>=2.2
numpy>=1.26
yfinance>=0.2.50
python-dateutil>=2.9
APScheduler>=3.10
jinja2>=3.1
diskcache>=5.6
tqdm>=4.66
# 선택 라이브러리 (원할 시 설치)
# alpha_vantage>=2.3
# sec-api>=1.0
# neo4j>=5.22
# ta>=0.11
# vectorbt>=0.26
# langchain-core>=0.3
# pyyaml>=6.0.2
# markdownify>=0.13
feedparser>=6.0.11
pyyaml>=6.0.2
fastmcp>=0.4.2
matplotlib>=3.9
